S:
I am trying to implement score_employees/4 to construct a list of doubles to which I can apply sort/4. However, it is not working as I expected. Doing Schedule/1 with this version outputs via writeln/1's what I expect in that the list is build, but it is returned as []

Further, I am building my list of doubles with a hardcoded "4", for now, but also tried to implement employee_Hours_In_Week/3 to get teh accurate list, but that caused an instant Abort, as it is currently coded.

T